Enhance retry delays with microsecond precision and configurable jitter
- Enable microsecond precision for retry delays using float values.
- Add configurable jitter percentages to strategies supporting jitter.
- Update documentation to clarify configuration options and base delay usage.
- Add comprehensive tests for delay calculations and jitter behavior.

+++ src/Contracts/RetryStrategy.php
@@ -10,10 +10,10 @@ interface RetryStrategy
      * Calculate the delay for the next retry attempt.
      *
      * @param  int  $attempt  Current attempt number (0-based)
-     * @param  float  $baseDelay  Base delay in seconds
-     * @return int Delay in seconds
+     * @param  float  $baseDelay  Base delay in seconds (can be float)
+     * @return float Delay in seconds (can have microsecond precision)
      */
-    public function getDelay(int $attempt, float $baseDelay): int;
+    public function getDelay(int $attempt, float $baseDelay): float;
 
     /**
      * Determine if another retry attempt should be made.
